refactor: move ping and disconnect tests from status manager
